+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/TimelineElementHolderViewModelTest.kt
@@ -658,6 +658,7 @@ class TimelineElementHolderViewModelTest {
             onMessageReport = mock(),
             onOpenMention = mock(),
             onOpenMetadata = mock(),
+            jumpTo = { _, _ -> }
         )
     }
 }